Finalização de PRojeto

+++ trabalho-final-web-ii/app/Http/Controllers/TodoGrupoController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Routing\Controller as Controller;
 use Illuminate\Support\Facades\DB;
 use App\Models\TodoGrupo;
+use App\Models\Todo;
 
 use Illuminate\Http\Request;
 
@@ -54,7 +55,16 @@ class TodoGrupoController extends Controller
     }
 
     public function remove($grupo_id){
-        TodoGrupo::where("id", $grupo_id)->delete();
-        return redirect(route("grupos.listagem"));
+        $todo = Todo::where("todo_groups_id", $grupo_id)->get()->first();
+
+        if(isset($todo->id) && !is_null($todo->id)){
+            return view('todo-grupos.listagem', [
+                'grupos' => TodoGrupo::get(),
+                'error'  => "Falha ao Remover Grupo. Há registros de To-Do relacionados ao grupo e portanto o mesmo não pode ser removido. Se realmente quiser remover este grupo, remova ou altere o grupo dos registros de To-Do conflitantes."
+            ]);
+        } else {
+            TodoGrupo::where("id", $grupo_id)->delete();
+            return redirect(route("grupos.listagem"));
+        }
     }
 }
